Several changes in EFAuthStore.cs
- Run GarbageCollect via System.Timers, make sure it runs every 15 minutes
- Permanent store via static variable, reduces amount of sql calls to database
- Narrower queries to database, search only for given key on RenewAsync, RetrieveAsync, RemoveAsync
- Handle exceptions in GarbageCollect 

If you want more background to these changes, we have previous code in production since a month back and we have application insights data and logs from errors generated by the previous code. We are also running the startup.cs with this authentication for different sitecore sites, so the static for timer and SQLAuthSessionStoreContext is partly due to that, although it could be handled in some other way, ie several different stores per site (that should be supported here by doing something with the constant for the connection string, should be set by startup.cs instead). 

For example we had the RetrieveAsync being executed tens of times per editor request, and in turn generating one sql query for all keys in database per execution, which amounted to something like 300 million sql queries perday in our prod env with like around 10 users average during the day, and one system user doing requests every 10 minutes. 

Then we also found out GarbageCollect wasn't running every 15 minutes as expected but every few seconds, this is either because we configured the store for several sites or because the previous timer was not static and was accumulating or a combination of both.  

Also, keys got impossible to unprotect in some environments, probably due to external circumstances to the SitecoreFederatedLogin module. Irregardless, this caused the log files be filled with exceptions errors (after we handled exceptions in GarbageCollect), aggravated by GarbageCollect being run every few seconds and being run once for each site.